Refactor; extract match fields definition so that it can be tied across ES1 and ES6 implementations.
