+++ media-api/app/lib/elasticsearch/MatchFields.scala
@@ -1,0 +1,17 @@
+package lib.elasticsearch
+
+import com.gu.mediaservice.lib.elasticsearch.ImageFields
+import lib.MediaApiConfig
+
+trait MatchFields extends ImageFields {
+
+  def config: MediaApiConfig
+
+  val matchFields: Seq[String] = Seq("id") ++
+    Seq("description", "title", "byline", "source", "credit", "keywords",
+      "subLocation", "city", "state", "country", "suppliersReference", "englishAnalysedCatchAll").map(metadataField) ++
+    Seq("labels").map(editsField) ++
+    config.queriableIdentifiers.map(identifierField) ++
+    Seq("restrictions").map(usageRightsField)
+
+}
