+++ media-api/app/lib/elasticsearch/impls/elasticsearch6/ElasticSearch.scala
@@ -16,7 +16,8 @@ import scalaz.syntax.std.list._
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class ElasticSearch(config: MediaApiConfig, mediaApiMetrics: MediaApiMetrics) extends ElasticSearchVersion with ElasticSearchClient with ImageFields {
+class ElasticSearch(val config: MediaApiConfig, mediaApiMetrics: MediaApiMetrics) extends ElasticSearchVersion with ElasticSearchClient
+  with ImageFields with MatchFields {
 
   lazy val imagesAlias = config.imagesAlias
   lazy val host = "localhost"
@@ -29,13 +30,6 @@ class ElasticSearch(config: MediaApiConfig, mediaApiMetrics: MediaApiMetrics) ex
 
   val searchFilters = new SearchFilters(config)
   val syndicationFilter = new SyndicationFilter(config)
-
-  val matchFields: Seq[String] = Seq("id") ++
-    Seq("description", "title", "byline", "source", "credit", "keywords",
-      "subLocation", "city", "state", "country", "suppliersReference", "englishAnalysedCatchAll").map(metadataField) ++
-    Seq("labels").map(editsField) ++
-    config.queriableIdentifiers.map(identifierField) ++
-    Seq("restrictions").map(usageRightsField)
 
   val queryBuilder = new QueryBuilder(matchFields)
